% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_factors.R
\name{summary.factor_select}
\alias{summary.factor_select}
\title{Summary Method for factor_select Objects}
\usage{
\method{summary}{factor_select}(object, ...)
}
\arguments{
\item{object}{A \code{factor_select} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{object}, called for its side effect
of printing a summary including leading eigenvalues to the console.
}
\description{
Summary Method for factor_select Objects
}
