% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_factors.R
\name{plot.factor_select}
\alias{plot.factor_select}
\title{Plot Method for factor_select Objects}
\usage{
\method{plot}{factor_select}(x, main = "Scree Plot", ...)
}
\arguments{
\item{x}{A \code{factor_select} object.}

\item{main}{Character string. Plot title. Defaults to \code{"Scree Plot"}.}

\item{...}{Further arguments passed to \code{plot()}.}
}
\description{
Produces a scree plot of the leading eigenvalues with the selected
number of factors marked.
}
